Partial fix for issue #173 (BEDES checkboxes not showing up for BEDES fields) There are three tables in the database that govern whether a field belongs to a schema (like BEDES):  seed_schema, seed_column, and seed_schema_columns. seed_schema has an id and a name.  Currently there is only one row with an id of 1 and a name of BEDES. seed_column has an id and a column_name field (it has more but these two are of interest here) seed_schema_columns is an index into seed_schema and seed_column identifying which schema has what columns.
For this fix I re-generated the list of bedes fields (bedes.py) to include the new fields.  It seems that you should just be able to re-run seed/migrations/0030_add_legit_bedes_columns.py again but figuring that out escapes me for the moment so I just made a new migration (0031...)

This does not fix the issue of why some fields are being marked as BEDES even though they are not. E.G. Release Date.  I did not see a Release Date field in the seed_schema_columns but have not really looked into that part of the bug yet.  Just putting this in because it seems to fix half the issue.