Separates org measurement prefs into import and display
Canadian municipalities and industry tend to use kWh/m2 for internal
discussion. Federal reporting and Portfolio Manager use GJ/m2. Two
things fall out of this:

- Choosing "metric" is not quite precise enough
- Just because incoming data is in GJ/m2 may not mean that the display
  units should be GJ/m2 (or vice-versa).

Therefore we should decouple these to allow an org to routinely bring in
GJ/m2 data without needing to commit to this as the display units.