Enables data quality checks for pint columns
Enable mins, maxes against pint columns with specified units. The rule
units don't have to match the base storage units, but this does
complicate the readable quality result string (ie. it's potentially a
bit confusing and needs tweaking).

Because this re-purposes the existing `units` column on the `seed_rule`,
there is not yet a dimensionality check: ie. has the user specified
`m**2` on an EUI-type column? This'll blow up in all sorts of
interesting ways.

Invalid picks should probably be blocked at the Angular level and have a
simple dimensionality check on the rule save API to throw back a 400 if
it's violated. Protect the boundary and then presume we'll be OK in the
check itself.