Fixes some obvious test failures
Not entirely clear to me whether this is the best way to handle these
but getting to green before refactor.

Main question here is whether the schema should report anything at all
for the pint columns given that they don't map based on the mapped name,
so for current purposes it's both moot and a bit ugly.