Adds angular translate, grunt extraction tasks
Need translate for on-the-fly translation in the UI, per Alex's
recommendation. The grunt tasks are to produce a translation file from
tagged UI strings, and (eventually) to convert back a .po file from
lokalize.co into a JSON file for consumption by angular-translate.

Adds a directory to keep locale files.