Introduces `translate` tags into Angular partials
Lets the grunt task auto-scan to extract them. Still need a strategy for
interpolation (mostly), hence the `TRANSLATION_FIXME` as a greppable
marker while I figure this out.

Haven't yet decided on what the dividing line is on deciding a chunk of
text needs to be replaced by a more generic key, but we'd need to
extract it into the translation file in any event and this accomplishes
that for later replacement in the partial.