Sets up interpolated translation, service translation
There are two situations in angular-translate needing some special
attention outside of just using the `<tag translate>` directive (or
`<tag>{$ 'my text' | translate $}</a>.

The first is where eg. Tax Lot or Property (or some variant) is being
injected into an otherwise static string based on a scope variable. I've
pulled these out so that it's a simple lookup with
`$scope.translate('THE_KEY')` and the ternary is otherwise untouched.
The downside is that we need separate translations for both the
consequent and the alternative, but this is *much* more direct and
understandable than trying to interpolate some variant (taxlot, tax lot,
tax lots, Tax Lot, Tax Lots). The list of strings to translate is just
assembled in the controller, piped through the `$translate` service and
parked on `$scope.translations`. We can see it all in one place. These
keys might have to be manually put into the locale files, but at least
they're not sprinkled all over the templates.

(I started with `translate-values="{
some_dict_key: massage(a_value) }" and this just got confusing and
overly vague and hard to proofread).

The second is a simple interpolation like "You have {$ count $} things."
and for this I think using `translate-values` is fine: there's no loss
of ability to proofread.