Ugh, trying out `sce` strategy
So. It is not going to be possible to avoid using the `{$ 'word' |
translate $}` form, which means we can't work around the double-escaping
problem from f1f7e045 (it happens regardless of whether we use unicode É
or `&Eacute;`. The `sce` strategy also does what we want at the cost of
failing the XSS test. Need to assess whether this is actually a problem.