Switch to messageformat interpolation
Pluralization in multiple languages is exactly the problem messageformat was
meant to solve, so this will take over from ng-pluralize, which can get
faintly ridiculous in the nested ternaries.

So, rather than `{$ $}` in the translation strings, it's simply `{ }`
(possibly with parameters for plural, eg `{var, plural, one{one
thing} few{a few things} other{# things}}`) and Lokalise is able to
syntax-color when it's correct.