Adds the display units to the column heads in ui-grid
Tack on the organization's display units before sending these out on the
column info API, and the ui-grids will pick it up. This is in keeping
with the general strategy of having Angular know nothing in particular
about the organization's choice of units and simply handling it on the
server side.

Still needs a converter from pint spec->display string.