Collapses quantities->floats before sending out inventory data
This can't easily be handled by a serializer, since the serializer needs
to be specialized by with some Org-level preferences. Adds a collapsing
function and maps it over the outbound property/taxlot dicts to convert
every Quantity object found to a straight float in the Org's preferred
units. Supercedes 0acd9a80.