Prettifies the ui-grid column headers and strips off `(Pint)`
When we swap the old columns out and the new Pint-aware columns in,
we'll not (presumably) need to strip the `(Pint)` off the tail end of
the display title. Even in the intermediate period, though, we haven't
lost the ability to tell which columns are Pint-aware, since this now
manifests in the UI as a column name with the units tacked on.