Respect units in report graphs
We convert the org's units same as in other places: by using
`apply_unit_display_preferences` in Python before shipping it out to
Angular.

Because units prefs are on org, we need to feed an org payload into the
reports controller. Note this just gives the right units on the
D3-rendered axes.

Wrapped the display in the same flipper as the rest of the UI until we
do the column swap of eg. gross_floor_area -> gross_floor_area_pint.