Migrates in new pint columns for old float columns
To reduce risk on making things units-aware, we created new columns for
the EUIs and areas that had the suffix `_pint`, and left the old ones in
place. This commit, via renaming, archives the old EUI/area columns with
the suffix `_orig`, and takes over the column names for the new
unit-aware Quantity columns.

This commit also provides a migration that simply copies the original
EUI/area data from the `_orig` Float columns to the corresponding new
Quantity column. This works because the implicit units of the
unit-unaware Float EUIs was `kBtu/ft**2/year` and these are the explicit
base units of the EUI Quantity columns. (Similarly for areas with
`ft**2`).  So the database representations are equivalent and any new
capabilities are being provided at the application/Django layer.

The columns technically could have been converted in place, but this
swap approach provides us a clean way to back out in case of problems:
remove the Quantity-based code and columns, and provide a migration that
removes the `_orig` suffix from the archived columns.