Converts existing and default rules to be Quantity-aware
Adds a migration that will update existing rules in place, changing over
the human-readable units string into one that Pint can use to do
Quantity-aware data checks. The Pint spec is also used in the `<select>`
in the rules table UI.

Modifies the definition of the default rules to provide valid Pint specs
in the standard US Customary base units for EUI, area.