Starts removing use_pint JS flipper, fixes mapping units
Revises the check code for whether we should show the units select when
setting up mappings. Separates the unit select into two with `ng-if` so
that it's only possible to select area units for area columns, EUI units
for EUI columns. Reverses the typeahead suggestion pruner to remove the
`*_orig` columns instead of the `*_pint` columns.