Remove use_pint flipper from reports controller
Collapse the X, Y axis choices to remove the _pint column choices.
Django is already handling collapsing down the units so the existing X,
Y axis choices became unit-aware after swapping the names in the
database.

Note that the area binning for the histogram becomes much less
interesting with square metres, but I'm considering making these to be
adaptive to the range of the data out-of-scope for now.