Ensures mappings have a unit selected if needed
Bad Things™ happen if we try to map a quantity field without giving a
unit in the mappings. This commit guards that by ensuring that the
`<select>` in the mappings table has a value because `from_units` on the
tcm always has a value. It's not yet hitched up to either org
preferences or (better) the `units_pint` stored in the `seed_column`
table to give back the previously-used value since this gets deep into
figuring out how to change the suggestions payload, which tbh I find a
bit hairy to work with.