Fixes logging error, quiets bad-data reporting
Without the off-by-one fix, `test_account_views.py` throws with a
tuple-outside-range error when it tries to log missing sig figs in the
setup payload.

Generally `None` is an acceptable datum here, so stop logging if we got
nothing at all. The intent of warning in the logs is to show that
someone is either deliberately or by mistake stuffing in bad (ie.
unparseable) data. *That* we should know about. Nothing at all isn't
really an actionable problem.