Fixes database error in `test_merge_duplicate_rows`
The tests for merge were failing on database errors: test setup was
attempting to stuff a `''` into the `gross_floor_area` column of a
PropertyState, which was failing because either a real double or `NULL`
is considered valid. It would only happen if the incoming Excel data had
an empty cell for a Quantity column, as
`example-data-properties-duplicates.xlsx` does.

This patch brute-force guards against these nils propagating past the
cleaner by effectively making `u''` a synonym for `None` and skipping
the cleaner entirely when it finds one (same as `NONE_SYNONYMS').

It appears the cleaner was not correctly identifying these as Pint (or
even float) in the ontology because the determination of what counts as
a quantity-type mapping is determined by the *raw* mapping. This is
because we could have multiple raw mappings to `gross_floor_area` but
with different units (eg. "Gross Floor Area (m2)" or "Gross Floor Area
(ft2)". The test suite fake mappings predate the Quantity fields, so
they don't designate an incoming unit. (So another fix for the symptom--
but not the actual problem--would be to provide `from_units` in
`seed/data_importer/tests/util.py#L99-101`).

Note that we didn't see the failure on the `site_eui` with that test
data because there are no empty cells in the Excel file mentioned above.
Things generally work out at the SQL level with old un-unitted mappings
(as we'd expect to see in production) because these get caught as
floats, and the cleaner's behaviour with floats is to either:

- to successfully parse (and therefore get stuffed into Postgres as the
  system's base units and be correctly interpreted as `ft**2` or
  `kBtu/ft**2/year`), or

- to fail to parse and then hand back `None` which Postgres is also OK
  with being `NULL`

It's not obvious how to migrate all existing Quantity `seed_columns` in
production databases to add units, but we could catch the obvious ones
(eg. Portfolio Manager default raw column names). I don't believe it's a
significant production issue though, as the very next time someone maps
data using the column its definition will get updated to tack on units;
the new mapping UI won't permit the unit to be blank.

Semantically, *explicitly* treating the empty string as `None` at the
cleaner feels like a better move than edge-casing the Pint columns
downstream in the `PropertyStateSerializer`. But, I'm more than open to
debate here since this affects every value, not just Pint values.

See also 2feb40d9. Part of #1536.