Proof-of-concept fix for 2.1 properties render failure
There is a bad interaction going on between [rest-framework][1] and
[pint][2]. The stock JSONEncoder used by rest-framework JSONRenderer
does a duck-type check to see if the thing being encoded responds to
`tolist()`. Sadly, a pint `Quantity` object does, and it'll throw an
exception if it's called with a singular object. It seems to exist for
convenience in dealing with numpy data frames to be able to assign a
unit to a frame or something list-like:

```py
>>> import pint
>>> ureg = pint.UnitRegistry()
>>> x = ureg.ft * 12.5
>>> x
<Quantity(12.5, 'foot')>
>>> x.tolist()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "pint/quantity.py", line 1370, in tolist
    for value in self._magnitude.tolist()]
AttributeError: 'float' object has no attribute 'tolist'
>>> y = ureg.ft * [12.5,13.5]
>>> y
<Quantity([ 12.5  13.5], 'foot')>
>>> y.tolist()
[<Quantity(12.5, 'foot')>, <Quantity(13.5, 'foot')>]
```

So `test_search_identifier` will fail on the get call to
`/api/v2.1/properties/` because the site_eui trips this codepath for
`tolist()` per the example above.

I can't really imagine how to change either rest-framework or pint to
reduce this interaction, so all that's left is to collapse Quantity
objects before it hits the renderer (same as we've been doing to
generally shield Angular from having to know anything at all about
Quantities).

This solution is incomplete; it will need to know about the API
consumer's organizational preferences to collapse to the right units and
sig-figs. The to-do here is to rely on the same machinery as the other
sites where we collapse (see also f99de48b). This commit is aimed just
at getting broad agreement that this is an acceptable course.

Part of #1536.

[1]: https://github.com/encode/django-rest-framework/blob/master/rest_framework/utils/encoders.py#L53-L55
[2]: https://github.com/hgrecco/pint/blob/master/pint/quantity.py#L1367-L1370