Fixes API test failure during mapping.
The all-the-way-through headless API test was failing on the mapping
step. The mappings are specified in a CSV fixture file. These predate
the Pint work and so didn't actually supply a `from_units` mapping,
which the Angular UI now always does. Without the `from_units`, the
GBA -> `gross_floor_area` mapping was throwing an exception because it
wasn't getting properly identified as a float. `map_row_chunk` got
fooled into thinking that "264,949" was a string, trying to stuff it
into a `double` database column (the underlying column type for Quantity
fields) and bubbling up an exception from Postgres.

This commit adds a proper `from_units` mapping to the fixture data, and
patches the request helper that the test uses for posting to
`/api/v2/import_files/1/save_column_mappings/` to make use of it.