Allows env to control postgres port in SEED docker image
Setting the postgresql port via an env var is consistent with the
other settings versus hard-coding. This allows the port used by a SEED
docker container to communicate with a postgres installation to be
re-mapped eg. in a docker-compose file. Our use case was to allow a
local dev installation of postgres to co-exist with one running in a
container.